Add: Management of Admin Modify Himself

+++ Controllers/Admin/DeleteUser.ts
@@ -4,6 +4,8 @@ import AdminService from "../../Services/AdminService";
 const deleteUser = async (req: Request, res: Response) => {
   try {
     const { id_delete_user } = req.params;
+    const { id_user } = req.body;
+    if (parseInt(id_delete_user) == parseInt(id_user)) { return res.status(403).json({ status: false, message: "No eliminar tu usuario." }); }
 
     const result: any = await AdminService.deleteUser(parseInt(id_delete_user));
     return res.status(result ? 200 : 400).json({
